% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermod_read_plt.R
\name{aermod_read_plt}
\alias{aermod_read_plt}
\title{Read AERMOD.PLT output file as tabular data}
\usage{
aermod_read_plt(
  path,
  cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC", ID = "NET ID")
)
}
\arguments{
\item{path}{a string to the *.PLT file}

\item{cols}{a character vector to select and optionally rename columns of the PLT file}
}
\value{
an object of type table
}
\description{
Return a table of the output file AERMOD.PLT
}
\examples{
\dontrun{
aermod_read_plt('./AERMOD.PLT')
}

}
