% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermod_stat_plt.R
\name{aermod_stat_plt}
\alias{aermod_stat_plt}
\title{Statistics of AERMOD PLT output file}
\usage{
aermod_stat_plt(plt, x_source, y_source)
}
\arguments{
\item{plt}{string for the aermod plt file}

\item{x_source}{numeric, x coordinate of the point source (or the centre or any point of the source)}

\item{y_source}{numeric, y coordinate of the point source (or the centre or any point of the source)}
}
\value{
a tibble with x and y coordinates, concentration values, distances from the source, compass degrees from the North corresponding to min and max statistics
}
\description{
Calculate statistics for AERMOD PLT output file
}
\examples{
\dontrun{
aermod_stat_plt('AERMOD.PLT', 739050, 5084550)
}
}
