% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_convert.R
\name{car2deg}
\alias{car2deg}
\title{Cartesian x, y to angle degrees}
\usage{
car2deg(x, y)
}
\arguments{
\item{x}{a numeric vector of x coordinates in the Cartesian plane}

\item{y}{a numeric vector of y coordinates in the Cartesian plane}
}
\value{
a numeric vector of corresponding angle degrees (anticlockwise direction)
}
\description{
Convert Cartesian coordinates x and y to angle degrees
}
\details{
By approaching the problem with the use of complex numbers (strange enough but it's easier!)
If z = x + i*y with real x and y
r = Mod(z) = sqrt(x^2 + y^2)
phi = Arg(z)
x = r * cos(phi)
y = r * sin(phi)
}
\examples{
x <- c(1, 1, 0, -1, -1, -1,  0,  1)
y <- c(0, 1, 1,  1,  0, -1, -1, -1)
car2deg(x, y)
}
