% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coord_dms_to_dd.R
\name{coord_dms_to_dd}
\alias{coord_dms_to_dd}
\title{coord_dms_to_dd}
\usage{
coord_dms_to_dd(s)
}
\arguments{
\item{s}{a vector of strings in the form "degrees minutes seconds" separated by a blank space}
}
\value{
a numeric vector in decimal degrees
}
\description{
Degrees minutes seconds to decimal degrees
}
\examples{
myangle<- "45  12    3.01"
coord_dms_to_dd(myangle)
}
