% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_convert.R
\name{deg2car}
\alias{deg2car}
\title{Angle degrees to Cartesian x, y coordinates}
\usage{
deg2car(deg, r = 1)
}
\arguments{
\item{deg}{a numeric vector of angle degrees}

\item{r}{numeric, ray of the unitary circle}
}
\value{
a tibble with degrees, phi in radians, x and y
}
\description{
Convert angle degrees to Cartesian coordinates x and y
It is internally also taking care of the conversion from input angles as degrees to radians
}
\examples{
deg<- c(0, 30, 45, 60, 90, 180, 270, 360)
deg2car(deg)
}
