% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/angle_convert.R
\name{deg2rad}
\alias{deg2rad}
\title{Angle degrees to radians}
\usage{
deg2rad(deg)
}
\arguments{
\item{deg}{a numeric vector of angles degrees}
}
\value{
a numeric vector of corresponding radians
}
\description{
Convert angles degrees to radians
}
\examples{
deg<- c(0, 30, 45, 60, 90, 180, 270, 360)
deg2rad(deg)
}
