% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degr_to_sect.R
\name{degr_to_sect}
\alias{degr_to_sect}
\title{wind dir convert degrees to sectors}
\usage{
degr_to_sect(vd)
}
\arguments{
\item{vd}{a numeric vector of wind dir degrees}
}
\value{
a character vector of wind dir sectors
}
\description{
wind dir convert degrees to sectors
}
\examples{
vd<-c(0, seq(11.25, (360-11.25), by = 22.5), 360)
degr_to_sect(vd)
}
