% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temp_convert.R
\name{F_convert}
\alias{F_convert}
\title{Temp conversion from Fahrenheit degrees F to ...}
\usage{
F_convert(deg = 0, to = "K")
}
\arguments{
\item{deg}{a numeric vector of Fahrenheit degrees}

\item{to}{an user input string for the conversion: "K" for Kelvin (default), "C" for Celsius degrees}
}
\value{
a numeric vector of the converted temperature degrees
}
\description{
Temp conversion from Fahrenheit degrees F to ...
}
\examples{
F_convert(0, "K")
F_convert(1, "C")
}
