% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mode.R
\name{get_mode}
\alias{get_mode}
\title{Get the mode}
\usage{
get_mode(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
get the mode (the most frequent value)
}
\description{
Get the mode
}
\examples{
v<- c(NA, 19, 4, 5, 7, 29, 19, 29, 13, 25, 19, 19)
mode(v)
}
