% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mypkg.R
\name{parse_version}
\alias{parse_version}
\title{parse_version}
\usage{
parse_version(ver)
}
\arguments{
\item{ver}{character string representing the version of a package}
}
\value{
an integer vector with the component of the version
}
\description{
parse the version of the installed package
}
\examples{
myRfuns:::parse_version('0.8')
}
