% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{the pipe operator}
\usage{
lhs \%>\% rhs
}
\description{
#' the code is quite exactly taken from https://github.com/UBESP-DCTV/imthcm
(thanks for that CL!)
}
\details{
See \code{magrittr::\link[magrittr]{\%>\%}} for details.
}
\keyword{internal}
