% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermod_read_plt.R
\name{plt_cols}
\alias{plt_cols}
\title{Internal function for selecting columns of AERMOD.PLT output file}
\usage{
plt_cols(
  path,
  cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC", ID = "NET ID"),
  skip = 6
)
}
\arguments{
\item{path}{a string for the path of PLT file}

\item{cols}{a character vector to select and optionally rename columns of the PLT file}

\item{skip}{number of rows to skip}
}
\value{
a tibble, identical in structure to the return value of \code{\link[readr:read_fwf]{readr::fwf_positions()}}
}
\description{
Return a column specification for the use with \code{\link[readr:read_fwf]{readr::read_fwf()}}.
}
