% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermod_plt_table.R
\name{read_plt}
\alias{read_plt}
\title{Import an AERMOD .PLT file as tabular data}
\usage{
read_plt(
  path,
  as = c("tbl", "tbl_cube", "array", "data.table", "raster"),
  cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC"),
  ...,
  verbose = getOption("verbose", default = FALSE)
)
}
\arguments{
\item{path}{(character) path to .PLT file(s)}

\item{as}{(character) format of result}

\item{verbose}{(logical)}
}
\value{
an object of type \code{as} (default: \code{tbl})
}
\description{
Import an AERMOD .PLT file as tabular data
}
