% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mypkg.R
\name{remote_version}
\alias{remote_version}
\title{remote_version}
\usage{
remote_version(con)
}
\arguments{
\item{con}{a connection to the remote DESCRIPTION file of the package}
}
\value{
a character string with the version of the package
}
\description{
get the version package in the remote repoitory of GitHub
}
\examples{

url(
  'https://raw.githubusercontent.com/maxbre/myRfuns/main/DESCRIPTION'
) \%>\%
myRfuns:::remote_version()

}
