% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_get_mode.R
\name{stat_get_mode}
\alias{stat_get_mode}
\title{stat_get_mode}
\usage{
stat_get_mode(v)
}
\arguments{
\item{v}{a numeric vector}
}
\value{
get the mode (the most frequent value)
}
\description{
get the mode of a numeric vector
}
\examples{
v<- c(NA, 19, 4, 5, 7, 29, 19, 29, 13, 25, 19, 19)
stat_get_mode(v)
}
