% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_mypkg.R
\name{update_mypkg}
\alias{update_mypkg}
\title{update_mypkg}
\usage{
update_mypkg()
}
\value{
invisible()
}
\description{
the aim of this function is to check and update the installed package
with any new version on GitHub
the code is quite exactly taken from https://github.com/UBESP-DCTV/imthcm
(thanks for that CL!)
}
\examples{
\dontrun{
    library(myRfuns)
    update_mypkg()
}
}
