% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_convert.R
\name{wind_degr_to_sect}
\alias{wind_degr_to_sect}
\title{wind_degr_to_sect}
\usage{
wind_degr_to_sect(vd)
}
\arguments{
\item{vd}{a numeric vector of wind dir degrees}
}
\value{
a character vector of wind dir sectors
}
\description{
convert wind degrees to sectors
}
\examples{
vd<-c(0, seq(11.25, (360-11.25), by = 22.5), 360)
wind_degr_to_sect(vd)
}
