% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wd_convert.R
\name{wind_sect_to_degr}
\alias{wind_sect_to_degr}
\title{wind_sect_to_degr}
\usage{
wind_sect_to_degr(vs)
}
\arguments{
\item{vs}{a character vector of wind dir sectors}
}
\value{
a numeric vector of wind dir degrees
}
\description{
convert wind sectors to degrees
}
\examples{
vs<-c("N","NNE","NE", "E", "ESE", "SE", "SSE", "S", "W", "WNW", "NW", "NNW")
wind_sect_to_degr(vs)
}
