% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermod_plt.R
\name{read_plt_tbl}
\alias{read_plt_tbl}
\title{Import AERMOD.PLT file as tabular data}
\usage{
read_plt_tbl(
  path,
  cols = c(X = "X", Y = "Y", CONC = "AVERAGE CONC", ID = "NET ID")
)
}
\arguments{
\item{path}{a string for the path of PLT file}

\item{cols}{a character vector to select and optionally rename columns of the PLT file}
}
\value{
an object of type tbl
}
\description{
Import AERMOD.PLT file as tabular data
}
\examples{
\dontrun{}

}
