% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aermod_map_plt_raster.R
\name{save_mapview_html}
\alias{save_mapview_html}
\title{save_mapview_html}
\usage{
save_mapview_html(my_mapview, string_filename_html)
}
\arguments{
\item{my_mapview}{a mapview object}

\item{string_filename_html}{a string to name the output html file}
}
\value{
a map in the form html and png
}
\description{
Convert mapview object to html and png files
}
