% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sect_to_degr.R
\name{sect_to_degr}
\alias{sect_to_degr}
\title{wind dir convert sectors to degrees}
\usage{
sect_to_degr(vs)
}
\arguments{
\item{vs}{a character vector of wind dir sectors}
}
\value{
a numeric vector of wind dir degrees
}
\description{
wind dir convert sectors to degrees
}
\examples{
vs<-c("N","NNE","NE", "E", "ESE", "SE", "SSE", "S", "W", "WNW", "NW", "NNW")
sect_to_degr(vs)
}
